ticketbuyer: write purchase info to db
writing purchase info to wstakemgr meta bucket for now.

+++ ticketbuyer/purchase.go
@@ -38,7 +38,7 @@ func (p *PurchaseManager) purchase(height int64) {
 		return
 	}
 	log.Debugf("Purchased %v tickets this round", purchaseInfo.Purchased)
-	p.w.SyncPurchased(height, purchaseInfo.Purchased)
+	p.w.SyncPurchased(height, int32(purchaseInfo.Purchased))
 }
 
 // NotificationHandler handles notifications, which trigger ticket purchases.
